reverted split
