--- app/client/src/ee/components/editorComponents/WidgetQueryGeneratorForm/CommonControls/DataSourceDropdown/useSource/utils.tsx
@@ -1,1 +1,0 @@
-export * from "ce/components/editorComponents/WidgetQueryGeneratorForm/commonControls/DatasourceDropdown/useSource/utils";
